Refactor file
