+++ gateway/src/model.rs
@@ -1,8 +1,6 @@
 use serde_derive::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use crate::compact_ids::CompactId;
-
 pub type GameId = Uuid;
 pub type ReqId = Uuid;
 pub type EventId = Uuid;
@@ -51,38 +49,6 @@ pub struct ProvideHistoryCommand {
     pub game_id: GameId,
     #[serde(rename = "reqId")]
     pub req_id: ReqId,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct ReconnectCommand {
-    #[serde(rename = "gameId")]
-    pub game_id: GameId,
-    #[serde(rename = "reqId")]
-    pub req_id: ReqId,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct JoinPrivateGameClientCommand {
-    #[serde(rename = "gameId")]
-    pub game_id: CompactId,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct FindPublicGameClientCommand {}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct CreatePrivateGameClientCommand {}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(tag = "type")]
-pub enum ClientCommands {
-    MakeMove(MakeMoveCommand),
-    Beep,
-    Reconnect(ReconnectCommand),
-    ProvideHistory(ProvideHistoryCommand),
-    JoinPrivateGame(JoinPrivateGameClientCommand),
-    FindPublicGame(FindPublicGameClientCommand),
-    CreatePrivateGame(CreatePrivateGameClientCommand),
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -166,47 +132,4 @@ pub struct Move {
     player: Player,
     coord: Option<Coord>,
     turn: i32,
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::compact_ids::CompactId;
-    use crate::model::*;
-    use uuid::Uuid;
-
-    #[test]
-    fn serialize_move_command() {
-        let game_id = Uuid::new_v4();
-        let req_id = Uuid::new_v4();
-
-        assert_eq!(
-            serde_json::to_string(&super::ClientCommands::MakeMove (super::MakeMoveCommand{
-                game_id,
-                req_id,
-                player: super::Player::BLACK,
-                coord: Some(super::Coord { x: 0, y: 0 })
-            }))
-            .unwrap(),
-            format!("{{\"type\":\"MakeMove\",\"gameId\":\"{:?}\",\"reqId\":\"{:?}\",\"player\":\"BLACK\",\"coord\":{{\"x\":0,\"y\":0}}}}", game_id, req_id)
-        )
-    }
-
-    #[test]
-    fn deserialize_join_priv_game_client_command() {
-        let compact_game_id = CompactId::encode(Uuid::new_v4());
-
-        let json = &format!(
-            "{{\"type\":\"JoinPrivateGame\",\"gameId\":\"{}\"}}",
-            compact_game_id.0
-        );
-
-        let d: ClientCommands = serde_json::from_str(json).unwrap();
-
-        assert_eq!(
-            d,
-            ClientCommands::JoinPrivateGame(JoinPrivateGameClientCommand {
-                game_id: compact_game_id
-            })
-        )
-    }
 }
